feat(receipt): attach PDF to confirmation emails and fix totals layout
Order confirmation emails now include the receipt PDF as a second
attachment alongside the existing ticket-and-QR-codes PDF. This covers
both the automatic flow (after payment/capture/mark-paid/free+invoice
checkout) and the manual "send email to buyer" action on the dashboard
order detail page. Buyers get tickets and receipt as two separate PDFs
in the same email.

Shared helper:
- src/lib/pdf/buildReceiptData.ts takes a Prisma client + order id and
  returns a ReceiptData payload ready for generateReceiptPdf. Returns
  null for orders in states that don't warrant a receipt (PENDING,
  CANCELLED). Used by both the download-receipt API route and the
  confirmation email so the two code paths stay in sync.

Email:
- sendOrderConfirmationEmail gains an optional orderId parameter. When
  present it calls buildReceiptDataForOrder + generateReceiptPdf and
  pushes a second attachment. Receipt generation failures are caught
  and logged so the buyer still gets their tickets if receipt build
  fails for any reason.
- All six call sites (checkout, capture, mark-paid, pay, dashboard
  email action, dashboard mark-paid action) now pass orderId.
- Email body copy switches between "your order confirmation is
  attached" and "your tickets and receipt are attached" depending on
  whether the receipt is present.

Receipt PDF formatting fixes (order #OE-MNRG91IM-N0SI):
- Widen the totals label column from 100pt to 190pt so long discount
  labels like "Discount (group 3+, 18.67% off)" fit on one line
  instead of wrapping and overlapping the next row. (Root cause of
  the overlap was that PDFKit's text(value, x, rowY) after a wrapped
  label reset doc.y back to one-line-below-rowY, so the next row
  started above the wrapped label.)
- Also make writeTotalsRow robust by advancing doc.y to
  max(labelBottom, valueBottom) before moveDown, so any future label
  that happens to wrap still doesn't overlap downstream rows.
- Replace the Unicode minus U+2212 in the discount row with an ASCII
  hyphen. PDFKit's default Helvetica uses WinAnsi encoding, which
  does not include U+2212, so it rendered as a stray "quote" glyph
  next to the amount.
