+++ src/app/api/orders/[id]/mark-paid/route.ts
@@ -198,6 +198,7 @@ export async function POST(request: NextRequest, context: RouteContext) {
     // Send confirmation email to buyer
     await sendOrderConfirmationEmail(paidOrder.buyerEmail, {
       orderNumber: paidOrder.orderNumber,
+      orderId: paidOrder.id,
       eventTitle: paidOrder.event.title,
       eventDate: formatDateTime(paidOrder.event.startDate),
       eventLocation:
