+++ src/app/api/orders/[id]/pay/route.ts
@@ -319,6 +319,7 @@ export async function POST(request: NextRequest, context: RouteContext) {
     try {
       await sendOrderConfirmationEmail(paidOrder.buyerEmail, {
         orderNumber: paidOrder.orderNumber,
+        orderId: paidOrder.id,
         eventTitle: paidOrder.event.title,
         eventDate: formatDateTime(paidOrder.event.startDate),
         eventLocation:
