+++ src/app/api/orders/by-number/[orderNumber]/receipt/route.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server'
 import { getCurrentUser } from '@/lib/auth'
 import { prisma } from '@/lib/db'
-import { generateReceiptPdf, type ReceiptData } from '@/lib/pdf/receipt'
+import { generateReceiptPdf } from '@/lib/pdf/receipt'
+import { buildReceiptDataForOrder } from '@/lib/pdf/buildReceiptData'
 
 export const dynamic = 'force-dynamic'
 
@@ -24,136 +25,19 @@ export async function GET(_request: Request, context: RouteContext) {
           OR: [{ userId: user.id }, { event: { organizer: { userId: user.id } } }],
         }),
       },
-      include: {
-        event: {
-          select: {
-            title: true,
-            startDate: true,
-            locationType: true,
-            venue: true,
-            city: true,
-            country: true,
-            onlineUrl: true,
-            organization: true,
-            organizationNumber: true,
-            organizationAddress: true,
-            organizer: {
-              select: {
-                orgName: true,
-                website: true,
-              },
-            },
-          },
-        },
-        items: {
-          include: {
-            ticketType: {
-              select: {
-                name: true,
-              },
-            },
-          },
-        },
-        groupDiscount: {
-          select: {
-            minQuantity: true,
-            discountType: true,
-            discountValue: true,
-          },
-        },
-        discountCode: {
-          select: {
-            code: true,
-            discountType: true,
-            discountValue: true,
-          },
-        },
-      },
+      select: { id: true, orderNumber: true, status: true },
     })
 
     if (!order) {
       return NextResponse.json({ error: 'Order not found' }, { status: 404 })
     }
 
-    // Don't issue receipts for orders that aren't financially settled in any way.
-    // PENDING (temporary pre-payment hold) and CANCELLED are excluded.
-    const allowedStatuses = ['PAID', 'PENDING_INVOICE', 'REFUNDED', 'PARTIALLY_REFUNDED']
-    if (!allowedStatuses.includes(order.status)) {
+    const receiptData = await buildReceiptDataForOrder(prisma, order.id)
+    if (!receiptData) {
       return NextResponse.json(
         { error: 'Receipt is not available for this order' },
         { status: 400 }
       )
-    }
-
-    const eventLocation =
-      order.event.locationType === 'ONLINE'
-        ? order.event.onlineUrl || 'Online event'
-        : [order.event.venue, order.event.city, order.event.country]
-            .filter(Boolean)
-            .join(', ') || 'Location TBD'
-
-    const buyerName = `${order.buyerFirstName} ${order.buyerLastName}`.trim()
-
-    const items = order.items.map((item) => {
-      const unitPrice = Number(item.unitPrice)
-      return {
-        name: item.ticketType.name,
-        quantity: item.quantity,
-        unitPrice,
-        lineTotal: Number(item.totalPrice),
-      }
-    })
-
-    let discountLabel: string | null = null
-    if (order.groupDiscount) {
-      const value = Number(order.groupDiscount.discountValue)
-      const valueLabel =
-        order.groupDiscount.discountType === 'PERCENTAGE'
-          ? `${value}%`
-          : `${value} ${order.currency}`
-      discountLabel = `group ${order.groupDiscount.minQuantity}+, ${valueLabel} off`
-    } else if (order.discountCode) {
-      discountLabel = `code ${order.discountCode.code}`
-    }
-
-    const receiptData: ReceiptData = {
-      orderNumber: order.orderNumber,
-      orderDate: order.createdAt,
-      paidAt: order.paidAt,
-      status: order.status,
-      paymentMethod: order.paymentMethod,
-      currency: order.currency,
-      seller: {
-        // Legal issuer name is per-event (falls back to the organizer's name
-        // for legacy events that predate the per-event issuer fields).
-        name: order.event.organization || order.event.organizer.orgName || 'Event Organizer',
-        displayName: null,
-        website: order.event.organizer.website,
-        orgNumber: order.event.organizationNumber,
-        address: order.event.organizationAddress,
-      },
-      buyer: {
-        name: buyerName,
-        email: order.buyerEmail,
-        title: order.buyerTitle,
-        organization: order.buyerOrganization,
-        address: order.buyerAddress,
-        city: order.buyerCity,
-        postalCode: order.buyerPostalCode,
-        country: order.buyerCountry,
-      },
-      event: {
-        title: order.event.title,
-        startDate: order.event.startDate,
-        location: eventLocation,
-      },
-      items,
-      subtotal: Number(order.subtotal),
-      discountAmount: Number(order.discountAmount),
-      discountLabel,
-      vatRate: Number(order.vatRate),
-      vatAmount: Number(order.vatAmount),
-      totalAmount: Number(order.totalAmount),
     }
 
     const pdfBuffer = await generateReceiptPdf(receiptData)
